+++ controllers/MaklumatKakitanganController.php
@@ -13,6 +13,8 @@ use yii\filters\VerbFilter;
 use yii\imagine\Image;
 use yii\web\UploadedFile;
 use kartik\mpdf\Pdf;
+use kartik\growl\Growl;
+
 /**
  * MaklumatKakitanganController implements the CRUD actions for MaklumatKakitangan model.
  */
@@ -72,25 +74,31 @@ class MaklumatKakitanganController extends Controller
         $model = new MaklumatKakitangan();
 
         if ($model->load(Yii::$app->request->post())) {
-            if (empty($model->file)) {
-                $model->foto = 'default-image.png';
-            }
-            else{
-                $model->file = UploadedFile::getInstance($model,'file');
+            $model->file = UploadedFile::getInstance($model,'file');
+            $nama = $model->nama;
+
+            if (!empty($model->file)) {
+                // $model->file = UploadedFile::getInstance($model,'file');
 
                 $model->file->saveAs('uploads/'.$model->file->baseName.'.'.$model->file->extension);
 
                 $model->foto = $model->file->baseName.'.'.$model->file->extension;
-
-                Image::thumbnail('@webroot/uploads/'.$model->foto, 300, 300)
-                        ->save(Yii::getAlias('@webroot/uploads/'.$model->foto), ['quality' => 90]);
+                
+            }
+            else{
+                
+                $model->foto = 'default-image.png';
+                // Image::thumbnail('@webroot/uploads/'.$model->foto, 300, 300)
+                //         ->save(Yii::getAlias('@webroot/uploads/'.$model->foto), ['quality' => 90]);
 
             }
             $model->status_pekerjaan = 1;
             
             $model->save();
 
-            Yii::$app->getSession()->setFlash('maklumatstaff-success', 'Tambah kakitangan berjaya');
+            $msg = 'Maklumat '.$nama.' Berjaya Di Simpan';
+            $this->getAddsuccess($msg);
+            // Yii::$app->getSession()->setFlash('maklumatstaff-success', 'Tambah kakitangan berjaya');
             return $this->redirect(['index']);
             // return $this->redirect(['view', 'id' => $model->id_staf]);
         } else {
@@ -111,6 +119,9 @@ class MaklumatKakitanganController extends Controller
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            $nama = $model->nama;
+            $msg = 'Kemaskini Maklumat '.$nama.' Berjaya';
+            $this->getAddsuccess($msg);
             return $this->redirect(['view', 'id' => $model->id_staf]);
         } else {
             return $this->render('_edit', [
@@ -119,6 +130,37 @@ class MaklumatKakitanganController extends Controller
         }
     }
 
+     public function actionUpdateimg($id)
+    {
+        $model = $this->findModel($id);
+
+        if ($model->load(Yii::$app->request->post())) {
+            $model->file = UploadedFile::getInstance($model,'file');
+            $nama = $model->nama;
+
+            if (!empty($model->file)) {
+                // $model->file = UploadedFile::getInstance($model,'file');
+
+                $model->file->saveAs('uploads/'.$model->file->baseName.'.'.$model->file->extension);
+
+                $model->foto = $model->file->baseName.'.'.$model->file->extension;
+                
+            }
+            else{
+                $model->foto = 'default-image.png';
+            }
+            
+            $msg = 'Kemaskini Gambar '.$nama.' Berjaya.';
+            $model->save();
+            $this->getAddsuccess($msg);
+
+            return $this->redirect(['view', 'id' => $model->id_staf]);
+        } else {
+            return $this->render('_edit', [
+                'model' => $model,
+            ]);
+        }
+    }
 
 
     /**
@@ -129,7 +171,21 @@ class MaklumatKakitanganController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        $model = $this->findModel($id);
+        // ->delete();Yii::getAlias('@webroot/uploads/'.
+        $pic = $model->foto;
+        $nama = $model->nama;
+        
+        if ($pic === 'default-image.png') {
+            $model->delete();
+            $this->getDeletemsg($nama);
+        }
+        else{
+            $model->delete();
+            unlink(Yii::getAlias('@webroot/uploads/'.$pic));
+            $this->getDeletemsg($nama);
+        }
+        
 
         return $this->redirect(['index']);
     }
@@ -402,5 +458,34 @@ class MaklumatKakitanganController extends Controller
             ]
         ]);
         return $pdf->render();
+    }
+
+    public function getDeletemsg($nama)
+    {
+        Yii::$app->getSession()->setFlash('success', [
+             'type' => Growl::TYPE_CUSTOM,
+             'duration' => 5000,
+             'icon' => 'glyphicon glyphicon-ok-sign',
+             'message' => 'Padam Maklumat '.$nama.' Berjaya',
+             'title' => 'Status',
+             'positonY' => 'top',
+             'positonX' => 'right',
+             'options'=>['class'=>'note note-success'],
+             // 'progressBarOptions' =>['class'=>'progress-bar progress-bar-success'],
+         ]);
+    }
+    public function getAddsuccess($msg)
+    {
+        Yii::$app->getSession()->setFlash('success', [
+             'type' => Growl::TYPE_CUSTOM,
+             'duration' => 5000,
+             'icon' => 'glyphicon glyphicon-ok-sign',
+             'message' => $msg,
+             'title' => 'Status',
+             'positonY' => 'top',
+             'positonX' => 'right',
+             'options'=>['class'=>'note note-success'],
+             // 'progressBarOptions' =>['class'=>'progress-bar progress-bar-success'],
+         ]);
     }
 }